refactor(studio-api): split agent route and service config internals
